requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
redis==5.0.1
psutil==5.9.6
python-dotenv==1.0.0
requests==2.31.0
eventlet==0.33.3